backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
cross-web==0.6.0
fastapi==0.135.3
graphql-core==3.2.8
graphql-ws==0.3.0
h11==0.16.0
idna==3.11
packaging==26.1
pydantic==2.13.1
pydantic_core==2.46.1
python-dateutil==2.9.0.post0
python-multipart==0.0.26
six==1.17.0
starlette==1.0.0
strawberry-graphql==0.314.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
websockets==16.0
yt-dlp==2026.3.17
